Relocate rgb->rgb_legacy
