+++ users/techworker/keycodes/process_records.c
@@ -8,11 +8,11 @@
 #include "version.h"
 
 #if defined(RGBLIGHT_ENABLE)
-#    include "rgb/rgb_stuff.h"
+#    include "rgb_legacy/rgb_stuff.h"
 #endif // defined(RGBLIGHT_ENABLE)
 
 #if defined(RGB_MATRIX_ENABLE)
-#    include "rgb/rgb_matrix_stuff.h"
+#    include "rgb_legacy/rgb_matrix_stuff.h"
 #endif // defined(RGB_MATRIX_ENABLE)
 
 #ifdef UNICODE_COMMON_ENABLE
