+++ users/techworker/rgb_legacy/anim/cool_diagonal.h
