+++ users/techworker/rgb_legacy/rgb.mk
@@ -1,9 +1,9 @@
 CUSTOM_RGBLIGHT ?= yes
 ifeq ($(strip $(RGBLIGHT_ENABLE)), yes)
     ifeq ($(strip $(CUSTOM_RGBLIGHT)), yes)
-        SRC += $(USER_PATH)/rgb/rgb_stuff.c
+        SRC += $(USER_PATH)/rgb_legacy/rgb_stuff.c
         OPT_DEFS += -DCUSTOM_RGBLIGHT
-        CONFIG_H += $(USER_PATH)/rgb/rgblight_config.h
+        CONFIG_H += $(USER_PATH)/rgb_legacy/rgblight_config.h
         ifeq ($(strip $(RGBLIGHT_NOEEPROM)), yes)
             OPT_DEFS += -DRGBLIGHT_NOEEPROM
         endif
@@ -16,9 +16,9 @@ endif
 CUSTOM_RGB_MATRIX ?= yes
 ifeq ($(strip $(RGB_MATRIX_ENABLE)), yes)
     ifeq ($(strip $(CUSTOM_RGB_MATRIX)), yes)
-        SRC += $(USER_PATH)/rgb/rgb_matrix_stuff.c
-        CONFIG_H += $(USER_PATH)/rgb/rgb_matrix_config.h
-        POST_CONFIG_H += $(USER_PATH)/rgb/post_rgb_matrix.h
+        SRC += $(USER_PATH)/rgb_legacy/rgb_matrix_stuff.c
+        CONFIG_H += $(USER_PATH)/rgb_legacy/rgb_matrix_config.h
+        POST_CONFIG_H += $(USER_PATH)/rgb_legacy/post_rgb_matrix.h
 
         OPT_DEFS += -DCUSTOM_RGB_MATRIX
         RGB_MATRIX_CUSTOM_USER = yes
