+++ users/techworker/rgb_matrix_user.inc
@@ -1,15 +1,15 @@
 // Copyright 2023 Christopher Courtney, aka Drashna Jael're  (@drashna) <drashna@live.com>
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-// #include "rgb/anim/startup_swirl_anim.h" // Requires matrix posistions to work
-// #include "rgb/anim/my_party_rocks.h" // seizure warning
-#include "rgb/anim/cool_diagonal.h"         // neaat
-#include "rgb/anim/random_breath_rainbow.h" // rgb vomit!
-// #include "rgb/anim/custom_gradient.h" // too custom
-#include "rgb/anim/raindrops.h"
-#include "rgb/anim/candy.h"
-// #include "rgb/anim/alpha_mod_homerow.h"
-// #include "rgb/anim/alpha_mod_homerow_cycle.h"
-// #include "rgb/anim/tetris_falling.h"
-// #include "rgb/anim/tetris_falling_horizontal.h"
-#include "rgb/anim/palette_fx.h"
+// #include "rgb_legacy/anim/startup_swirl_anim.h" // Requires matrix posistions to work
+// #include "rgb_legacy/anim/my_party_rocks.h" // seizure warning
+#include "rgb_legacy/anim/cool_diagonal.h"         // neaat
+#include "rgb_legacy/anim/random_breath_rainbow.h" // rgb vomit!
+// #include "rgb_legacy/anim/custom_gradient.h" // too custom
+#include "rgb_legacy/anim/raindrops.h"
+#include "rgb_legacy/anim/candy.h"
+// #include "rgb_legacy/anim/alpha_mod_homerow.h"
+// #include "rgb_legacy/anim/alpha_mod_homerow_cycle.h"
+// #include "rgb_legacy/anim/tetris_falling.h"
+// #include "rgb_legacy/anim/tetris_falling_horizontal.h"
+#include "rgb_legacy/anim/palette_fx.h"
