+++ users/techworker/rules.mk
@@ -73,7 +73,7 @@ endif
 
 # include $(USER_PATH)/pointing/pointing.mk
 include $(USER_PATH)/split/split.mk
-include $(USER_PATH)/rgb/rgb.mk
+include $(USER_PATH)/rgb_legacy/rgb.mk
 include $(USER_PATH)/keycodes/keycodes.mk
 # include $(USER_PATH)/features/common.mk
 # Ignore if not found
